Modified: Functions.

+++ app/Http/Controllers/BookController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Support\Facades\Cache;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Storage;
 use Illuminate\Http\Request;
 use App\Models\User;
 use App\Models\Book;
@@ -21,7 +22,7 @@ class BookController extends Controller
             'title' => 'required|string|between:2,50',
             'description' => 'required|string|between:3,1000',
             'author' => 'required|string|between:3,100',
-            'logo' => 'required|string|between:3,1000',
+            'image' => 'required|image|mimes:jpeg,png,jpg,gif,svg|max:2048',
             'price' => 'required|numeric',
             'quantity'=>'required|integer|min:1'
         ]);
@@ -47,11 +48,16 @@ class BookController extends Controller
                 {
                     return response()->json(['message' => 'You are trying to add Existing Book'], 401);
                 }
+
+                $imageName = time().'.'.$request->image->extension();  
+                $path = Storage::disk('s3')->put('images', $request->image);
+                $pathurl = Storage::disk('s3')->url($path);
+
                 $book = new Book;
                 $book->title = $request->input('title');
                 $book->description = $request->input('description');
                 $book->author = $request->input('author');
-                $book->logo = $request->input('logo');
+                $book->image = $pathurl;
                 $book->price = $request->input('price');
                 $book->quantity = $request->input('quantity');
                 $book->user_id = $currentUser->id;
@@ -60,11 +66,11 @@ class BookController extends Controller
         } 
 		catch (Exception $e) 
 		{
-            Log::error('Invalid User');
-            return response()->json([ 'message' => 'Invalid authorization token'], 404);
+             Log::error('Invalid User');
+             return response()->json([ 'message' => 'Invalid authorization token'], 404);
         }
 
-        Log::info('book created',['admin_id'=>$book->admin_id]);
+        Log::info('book created',['admin_id'=>$book->user_id]);
         return response()->json(['message' => 'Book created successfully'],201);
     }
 
@@ -93,6 +99,7 @@ class BookController extends Controller
             {
                 return response()->json(['message' => 'Could not found Book with that id'], 404);
             }
+
             $book->quantity += $request->quantity;
             $book->save();
             return response()->json([ 'message' => 'Book Quantity updated Successfully'], 201);
@@ -104,6 +111,7 @@ class BookController extends Controller
         }
     }
 
+
     public function updateBookById(Request $request)
     {
         $validator = Validator::make($request->all(), [
@@ -111,38 +119,56 @@ class BookController extends Controller
             'title' => 'string|between:2,50',
             'description' => 'string|between:3,1000',
             'author' => 'string|between:3,100',
-            'logo' => 'string|between:3,1000',
+            'image' => 'image|mimes:jpeg,png,jpg,gif,svg|max:2048',
             'price' => 'numeric',
             'quantity'=>'integer|min:1'
         ]);
+
         if($validator->fails())
         {
             return response()->json($validator->errors()->toJson(), 400);
         }
-        try
+        
+        try 
         {
-            $currentUser = JWTAuth::parseToken()->authenticate();
-            $admin = User::select('id')->where([
-                ['role','=','admin'],
-                ['id','=',$currentUser->id]
-            ])->get();
-
-            if(count($admin) == 0)
-            {
-                return response()->json(['message' => 'Un authorized' ], 401);
-            }
             $id = $request->input('id');
-            $book=Book::find($request->id);
-            if(!$book)
+            $currentUser = JWTAuth::parseToken()->authenticate();
+            if ($currentUser)
             {
-                return response()->json([ 'message' => 'Book not Found'], 404);
+                $admin = User::select('id')->where([
+                    ['role','=','admin'],
+                    ['id','=',$currentUser->id]
+                ])->get();
+        
+                if(count($admin)==0)
+                {
+                    return response()->json(['message' => 'Unauthorised'], 403);
+                }
+                $book = Book::find($request->id);
+                //echo $book;
+                if(!$book)
+                {
+                    return response()->json(['message' => 'Book not Found'], 404);
+                }
+
+                if($request->image)
+                {
+                    $path = str_replace(env('AWS_URL_PATH'),'',$book->image);
+            
+                    if(Storage::disk('s3')->exists($path)) {
+                        Storage::disk('s3')->delete($path);
+                    }
+                    $path = Storage::disk('s3')->put('images', $request->image);
+                    $pathurl = Storage::disk('s3')->url($path);
+                    $book->logo = $pathurl;
+                }
+
+                $book->fill($request->except('image'));
+                if($book->save())
+                {
+                    return response()->json(['message' => 'Book updated Sucessfully' ], 201);
+                }
             }
-            $book->fill($request->all());
-            if($book->save())
-            {
-                return response()->json(['message' => 'Book updated Sucessfully' ], 201);
-            } 
-                 
         }
         catch(Exception $e)
         {
@@ -172,17 +198,24 @@ class BookController extends Controller
             {
                 return response()->json(['message' => 'Un authorized' ], 401);
             }
-            $id = $request->input('id');
+            
             $book=Book::find($request->id);
             if(!$book)
             {
                 return response()->json([ 'message' => 'Book not Found'], 404);
             }
-            if($book->delete())
+
+            $path = str_replace(env('AWS_URL_PATH'),'',$book->image);
+            if(Storage::disk('s3')->exists($path)) 
             {
-                Log::info('book deleted',['user_id'=>$currentUser,'book_id'=>$request->id]);
-                return response()->json(['message' => 'Book deleted Sucessfully'], 201);
-            }    
+                Storage::disk('s3')->delete($path);
+                if($book->delete())
+                {
+                    Log::info('book deleted',['user_id'=>$currentUser,'book_id'=>$request->id]);
+                    return response()->json(['message' => 'Book deleted Sucessfully'], 201);
+                }
+            }     
+            return response()->json(['message' => 'File image was not deleted'], 402);    
         }
         catch(Exception $e)
         {
@@ -194,7 +227,7 @@ class BookController extends Controller
 
     public function getAllBooks()
     {           
-        $books = Book::select('id','title','description','author','logo','price','quantity')->get();
+        $books = Book::select('id','title','description','author','image','price','quantity')->get();
         if ($books=='[]')
         {
             return response()->json(['message' => 'Books not found'], 404);
